三つのテーブルの、カラムの指定、Eloquent ORMの設定、親子関係は
