+++ src/app/Http/Controllers/Auth/LoginController.php
@@ -37,4 +37,8 @@ class LoginController extends Controller
     {
         $this->middleware('guest')->except('logout');
     }
+    public function username()
+    {
+      return 'name';
+    }
 }
